requirements.txt
Django>=3.1, <3.3
python-decouple
djangorestframework
PyJWT==1.7.1
djangorestframework-simplejwt
celery>=4.4, <4.5

psycopg2_binary
# requests
# python-memcached
# Pillow

django-filter
drf-yasg
django-seed